+++ src/main/java/org/example/oddventure/domain/ai/controller/AiController.java
@@ -1,10 +1,10 @@
-package org.example.oddventure.domain.winningRateAi.controller;
+package org.example.oddventure.domain.ai.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.example.oddventure.common.dto.response.ApiResponse;
-import org.example.oddventure.domain.winningRateAi.dto.AiRequest;
-import org.example.oddventure.domain.winningRateAi.dto.AiResponse;
-import org.example.oddventure.domain.winningRateAi.service.AiService;
+import org.example.oddventure.domain.ai.dto.AiRequest;
+import org.example.oddventure.domain.ai.dto.AiResponse;
+import org.example.oddventure.domain.ai.service.AiService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,6 +18,6 @@ public class AiController {
     @GetMapping("/winningrate")
     ResponseEntity<ApiResponse<AiResponse>> generation(@RequestBody String content) {
         AiRequest request = new AiRequest(content);
-        return ApiResponse.success(aiService.generateAbnormalBehaviorReport(request), "ai 기반 승률 예측 결과를 반환한다.");
+        return ApiResponse.success(aiService.calculateWinningRateWithAi(request), "ai 기반 승률 예측 결과를 반환한다.");
     }
 }
